approval-service/requirements.txt
fastapi
uvicorn
requests
pika
sqlalchemy
pydantic
psycopg2-binary
python-dotenv